Another warning.

+++ Classes/PropertyView/SBChapterViewController.m
@@ -108,7 +108,7 @@
 
 - (IBAction)renameChapters:(id)sender {
     [track.chapters enumerateObjectsUsingBlock:^(id obj, NSUInteger idx, BOOL *stop) {
-        NSString *title = [NSString stringWithFormat:@"Chapter %lu", idx + 1];
+        NSString *title = [NSString stringWithFormat:@"Chapter %lu", (unsigned long) idx + 1];
         [(MP42TextSample *)obj setTitle:title];
     }];
